Added all.sh for running all scripts
Plus some tidying up.

+++ 6_vim.sh
@@ -1,4 +1,7 @@
 #!/bin/sh
 
-cp $HOME/.vimrc $HOME/.vimrc_old
+if [ -f $HOME/.vimrc ]
+then
+	cp $HOME/.vimrc $HOME/.vimrc_old
+fi
 cp .vimrc $HOME/
